ARTEMIS-2221 avoid unnecessary Bindings instance creation
When trying to get the bindings for an address the getBindingsForAddress
method will create a Bindings instance if there are no bindings for the
address. This is unnecessary in most circumstances so use the
lookupBindingsForAddress method instead and check for null.

+++ artemis-protocols/artemis-openwire-protocol/src/main/java/org/apache/activemq/artemis/core/protocol/openwire/OpenWireConnection.java
@@ -1033,17 +1033,19 @@ public class OpenWireConnection extends AbstractRemotingConnection implements Se
 
 
       } else {
-         Bindings bindings = server.getPostOffice().getBindingsForAddress(new SimpleString(dest.getPhysicalName()));
+         Bindings bindings = server.getPostOffice().lookupBindingsForAddress(new SimpleString(dest.getPhysicalName()));
 
-         for (Binding binding : bindings.getBindings()) {
-            Queue b = (Queue) binding.getBindable();
-            if (b.getConsumerCount() > 0) {
-               throw new Exception("Destination still has an active subscription: " + dest.getPhysicalName());
+         if (bindings != null) {
+            for (Binding binding : bindings.getBindings()) {
+               Queue b = (Queue) binding.getBindable();
+               if (b.getConsumerCount() > 0) {
+                  throw new Exception("Destination still has an active subscription: " + dest.getPhysicalName());
+               }
+               if (b.isDurable()) {
+                  throw new Exception("Destination still has durable subscription: " + dest.getPhysicalName());
+               }
+               b.deleteQueue();
             }
-            if (b.isDurable()) {
-               throw new Exception("Destination still has durable subscription: " + dest.getPhysicalName());
-            }
-            b.deleteQueue();
          }
       }
